requirements.txt
flask
anthropic
requests
python-dotenv
selenium
webdriver-manager

